Tutorial work
This is the start of the work on the Generated maze, I have followed a tutorial to get to this point I just need to fix the timer and score. After this I am going to use the stuff we learnt in class to add some more colour to the maze and maybe even some creatures or a VR rig.